Add skeleton of the initial feature spec
I want this spec to drive the implementation from the outside. The goal
is to be able to pass in a seed and predict the moves.

I added a basic interaction to experiment on how to do that from the
test.

STDIN and STDOUT needs to be synchronous so we don't have to deal with
buffering.